Added a common interface for definitions and values: Metadata
When we create values from tree node definitions, we will
have to compare definition and value trees, this requires
similar operations on both definitions and values.
To facilitate this, I have made a common interface for
values and definitions and renamed the parent interface to
MetaDataParent so that TreeNodeDefinition can use that too.

I have also added an interface for Values and one for Definitions,
MetadataValue and MetadataDefinition.

The added methods in ParentUtilTest has been copy pasted from
TreeStructureUtil since I am not sure yet if they are needed
for Definitions as well. If we find out that they are, they should
be added to TreeStructureUtil.

Change-Id: I70207120c164b1bfb4a56c71f5660694e8686254