ParentUtil can now do replace.
This was a tricky thing to get done, since the original algorithm
explicitly hindered stuff to be overwritten the replace algorithm
was added by itself.
The idea is that a standard TreeNode doesn't need to be replaced
since it is just a list, but some other kind of more complex parent
might need to, so they can signal that via a new method. And when a
value has been replaced the new object gets signalled that it a
replacement so that it can try to merge any old data if it wants to.

In order to be able to build with maven 3 I had to uppgrade the
checkstyle version which introduced new SWE checks that the old
version didn't look for. So if you see four or five lines changed
in a file it is probably just some import fixes for checkstyle.

Change-Id: I2941409020e6cf5f1fdb7ab2c43496f8c1fa930b