Added environment variable support for metadata values
Added a checkbox for each metadata value (and for string
metadata definition). When the build starts, if this check box
is ticked, the value will be converted to an environment variable.
Trees have a checkbox for adding all its children as variables.

Definition support was started but put on halt since it is not
prioritized.

Change-Id: I0ffe7164ef7a09638fcad7df3f9d11f209afdb71